split all /bot commands and /me triggers into plugin files
